+++ compiler/OutputToFont.cpp
@@ -910,7 +910,7 @@ bool GrcManager::AddFeatsModFamilyAux(uint8 * pTblOld, uint32 /*cbTblOld*/,
 	uint8 * pTblNew, uint32 /*cbTblNew*/,
 	std::vector<std::wstring> & vstuExtNames, std::vector<uint16> & vnLangIds,
 	std::vector<uint16> & vnNameTblIds, 
-	uint16 * pchwFamilyName, uint16 cchwFamilyName, std::vector<PlatEncChange> & vpec,
+	uint16 * pchwFamilyName, size_t cchwFamilyName, std::vector<PlatEncChange> & vpec,
 	int nNameTblMinNew)
 {
     #pragma pack(1)
